Adds the mouse dragging to pan function
This function is implemented in onRender() instead of onMouseMove(),
because it will be smoother/responsive this way.

Strangely, on the other hand, I found out that the mouse wheel to zoom
function is actually smoother if implemented in the onMouseWheel().
If implemented in onRender(), wheel events will sometimes be ignored.
I cannot explain why.